Use schema parser.
